requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.11.8
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
djongo==1.3.3
Markdown==3.3.3
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pymongo==3.11.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
sqlparse==0.2.4
wincertstore==0.2
